requirements.txt
aiohttp==3.9.5
Flask==3.0.3
inquirer==3.3.0
parse_torrent_title==2.8.1
PlexAPI==4.15.15
questionary==2.0.1
Requests==2.32.3
rich==13.7.1
tabulate==0.9.0
tqdm==4.66.4
urwid==2.6.15
pytrakt==3.4.32
flask_session==0.8.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
bencodepy==0.9.5
flask-login==0.6.3
flask-sqlalchemy==3.1.1
unidecode==1.3.8
babelfish==0.6.1 
guessit==3.8.0 
rebulk==3.2.0
jaconv==0.4.0 
pykakasi==2.3.0
tenacity==9.0.0
iso8601==2.1.0
